feat(Task): 할 일 생성 기능 구현
1. 할 일 생성하는 기능 구현
2. 카테고리 레포지토리 구현

+++ src/main/java/ToDo/example/domain/Category.java
@@ -11,11 +11,12 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int categoryId;
+    private Long categoryId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
+
     private String categoryName;
     private String categoryColor;
 }
